historical-data-service/requirements.txt
grpcio
grpcio-tools
redis
redis-om